Implementação de transação para exclusão de dados utilizando Unit Of Work.

+++ ProjetoEscola.API/Controllers/SubjectsController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Mvc;
 using ProjetoEscola.Application.DTO_s;
 using ProjetoEscola.Application.Services.Interfaces;
+using ProjetoEscola.Domain.Interface;
 
 namespace ProjetoEscola.API.Controllers
 {
@@ -11,13 +12,15 @@ namespace ProjetoEscola.API.Controllers
         private readonly ISubjectService _subjectService;
         private readonly IStudentsSubjectsService _studentsSubjectsService;
         private readonly ITeacherSubjectService _teacherSubjectService;
+        private readonly IUnitOfWork _unitOfWork;
 
 
-        public SubjectsController(ISubjectService subjectService, IStudentsSubjectsService studentsSubjectsService, ITeacherSubjectService teacherSubjectService)
+        public SubjectsController(ISubjectService subjectService, IStudentsSubjectsService studentsSubjectsService, ITeacherSubjectService teacherSubjectService, IUnitOfWork unitOfWork)
         {
             _subjectService = subjectService;
             _studentsSubjectsService = studentsSubjectsService;
             _teacherSubjectService = teacherSubjectService;
+            _unitOfWork = unitOfWork;
         }
 
         /// <summary>
@@ -107,20 +110,34 @@ namespace ProjetoEscola.API.Controllers
         [HttpDelete]
         public async Task<ActionResult> DeleteAsync(int id)
         {
+            await _unitOfWork.BeginTransactionAsync ();
             try
             {
                 var studentSuject = await _studentsSubjectsService.GetBySubjectIdAsync(id);
-                if(studentSuject.Data != null)
+                if (studentSuject.Data != null)
                 {
                     await _studentsSubjectsService.DeleteAsync(studentSuject.Data.Id);
                 }
 
-                var result = await _subjectService.DeleteAsync(id);
+                var teacherSubject = await _teacherSubjectService.GetBySubjectIdAsync(id);
+                if (teacherSubject.Data != null)
+                {
+                    await _teacherSubjectService.DeleteAsync(teacherSubject.Data.Id);
+                }
 
-                return BadRequest(result);
+                var subject = await _subjectService.GetByIdAsync(id);
+                if (subject.Data != null)
+                {
+                    await _subjectService.DeleteAsync(id);
+                }
+
+                await _unitOfWork.CommitAsync();
+
+                return Ok(subject);
             }
             catch (Exception ex)
             {
+                await _unitOfWork.RollbackAsync();
                 throw new Exception(ex.Message);
             }
         }
